Refactor commands to accomodate config
