+++ src/main/java/seedu/todolist/constants/Formats.java
@@ -9,8 +9,8 @@ public final class Formats {
     public static final String TIME_IN_1 = "[d-M-uuuu H:m]";
     public static final String TIME_IN_2 = "[d/M/uuuu H:m]";
     public static final String TIME_OUT = "dd LLL uuuu HH:mm";
-    public static final String CONFIG_INFO = "Repeating tasks every %d days\nChecking for repeating task every: " +
-                                                   "%d minutes\nLast checked at: %s";
+    public static final String CONFIG_STRING = "Repeating tasks every %d days" + System.lineSeparator()
+        + "Checking for repeating task every: %d minutes" + System.lineSeparator() + "Last checked at: %s";
 
     public static final DateTimeFormatter TIME_IN_FORMATTER = DateTimeFormatter
             .ofPattern(Formats.TIME_IN_1 + Formats.TIME_IN_2).withResolverStyle(ResolverStyle.STRICT);
