+++ src/main/java/seedu/todolist/logic/command/AddTaskCommand.java
@@ -3,8 +3,9 @@ package seedu.todolist.logic.command;
 import seedu.todolist.exception.ToDoListException;
 import seedu.todolist.constants.Flags;
 import seedu.todolist.logic.ParserUtil;
+import seedu.todolist.model.Config;
 import seedu.todolist.ui.Ui;
-import seedu.todolist.task.TaskList;
+import seedu.todolist.model.TaskList;
 
 import java.time.LocalDateTime;
 import java.util.HashMap;
@@ -42,7 +43,7 @@ public class AddTaskCommand extends Command {
     }
 
     @Override
-    public void execute(TaskList taskList, Ui ui) {
+    public void execute(TaskList taskList, Config config, Ui ui) {
         String taskString = taskList.addTask(description, deadline, email, tags, repeatDuration, priority);
         ui.printAddTaskMessage(taskString);
     }
