+++ src/main/java/seedu/todolist/logic/command/Command.java
@@ -1,8 +1,9 @@
 package seedu.todolist.logic.command;
 
 import seedu.todolist.exception.ToDoListException;
+import seedu.todolist.model.Config;
 import seedu.todolist.ui.Ui;
-import seedu.todolist.task.TaskList;
+import seedu.todolist.model.TaskList;
 
 /**
  * Parent class for Command objects.
@@ -11,7 +12,7 @@ public abstract class Command {
     /**
      * Executes the command.
      */
-    public void execute(TaskList taskList, Ui ui) throws ToDoListException{}
+    public abstract void execute(TaskList taskList, Config config, Ui ui) throws ToDoListException;
 
     /**
      * Returns whether a command, by default, should terminate the program.
