+++ src/main/java/seedu/todolist/logic/command/EditDeadlineCommand.java
@@ -7,7 +7,8 @@ import seedu.todolist.exception.InvalidIdException;
 import seedu.todolist.exception.ToDoListException;
 import seedu.todolist.logic.FormatterUtil;
 import seedu.todolist.logic.ParserUtil;
-import seedu.todolist.task.TaskList;
+import seedu.todolist.model.Config;
+import seedu.todolist.model.TaskList;
 import seedu.todolist.ui.Ui;
 
 import java.time.LocalDateTime;
@@ -32,7 +33,7 @@ public class EditDeadlineCommand extends Command  {
     }
 
     @Override
-    public void execute(TaskList taskList, Ui ui) throws InvalidIdException {
+    public void execute(TaskList taskList, Config config, Ui ui) throws InvalidIdException {
         for (int id : idHashSet) {
             String taskString = taskList.setDeadline(id, deadline);
             if (deadline == null) {
