+++ src/main/java/seedu/todolist/logic/command/EditDescriptionCommand.java
@@ -4,7 +4,8 @@ import seedu.todolist.constants.Flags;
 import seedu.todolist.exception.InvalidIdException;
 import seedu.todolist.exception.ToDoListException;
 import seedu.todolist.logic.ParserUtil;
-import seedu.todolist.task.TaskList;
+import seedu.todolist.model.Config;
+import seedu.todolist.model.TaskList;
 import seedu.todolist.ui.Ui;
 
 import java.util.HashMap;
@@ -24,7 +25,7 @@ public class EditDescriptionCommand extends Command {
     }
 
     @Override
-    public void execute(TaskList taskList, Ui ui) throws InvalidIdException {
+    public void execute(TaskList taskList, Config config, Ui ui) throws InvalidIdException {
         for (int id: idHashSet) {
             String taskString = taskList.setDescription(id, description);
             ui.printEditTaskMessage("description", description, taskString);
