+++ src/main/java/seedu/todolist/logic/command/ExitCommand.java
@@ -1,7 +1,8 @@
 package seedu.todolist.logic.command;
 
+import seedu.todolist.model.Config;
 import seedu.todolist.ui.Ui;
-import seedu.todolist.task.TaskList;
+import seedu.todolist.model.TaskList;
 
 /**
  * Command for terminating the program.
@@ -10,7 +11,7 @@ public class ExitCommand extends Command {
     /**
      * Displays the program's shutdown message.
      */
-    public void execute(TaskList taskList, Ui ui) {
+    public void execute(TaskList taskList, Config config, Ui ui) {
         ui.printGoodbyeMessage();
     }
