+++ src/main/java/seedu/todolist/logic/command/FindByTagCommand.java
@@ -3,8 +3,9 @@ package seedu.todolist.logic.command;
 import seedu.todolist.constants.Flags;
 import seedu.todolist.exception.InvalidFindException;
 import seedu.todolist.exception.ToDoListException;
-import seedu.todolist.task.Task;
-import seedu.todolist.task.TaskList;
+import seedu.todolist.model.Config;
+import seedu.todolist.model.Task;
+import seedu.todolist.model.TaskList;
 import seedu.todolist.ui.Ui;
 
 import java.util.ArrayList;
@@ -23,7 +24,7 @@ public class FindByTagCommand extends Command{
         }
     }
     @Override
-    public void execute(TaskList taskList, Ui ui) throws InvalidFindException {
+    public void execute(TaskList taskList, Config config, Ui ui) throws InvalidFindException {
         if (taskList.getAllTags().contains(tag)) {
             ArrayList<Task> arrayList = taskList.getTaskWithTag(tag);
             ui.printTasksWithTag(arrayList);
