+++ src/main/java/seedu/todolist/logic/command/ListFullInfoCommand.java
@@ -5,7 +5,8 @@ package seedu.todolist.logic.command;
 import seedu.todolist.constants.Flags;
 import seedu.todolist.exception.InvalidIdException;
 import seedu.todolist.logic.ParserUtil;
-import seedu.todolist.task.TaskList;
+import seedu.todolist.model.Config;
+import seedu.todolist.model.TaskList;
 import seedu.todolist.ui.Ui;
 
 import java.util.HashMap;
@@ -22,7 +23,7 @@ public class ListFullInfoCommand extends Command{
     }
 
     @Override
-    public void execute(TaskList taskList, Ui ui) throws InvalidIdException {
+    public void execute(TaskList taskList, Config config, Ui ui) throws InvalidIdException {
         for (int id : idHashSet) {
             ui.printGetFullInfoMessage(taskList.getFullInfo(id));
         }
