+++ src/main/java/seedu/todolist/logic/command/UnmarkTaskCommand.java
@@ -4,7 +4,8 @@ package seedu.todolist.logic.command;
 import seedu.todolist.constants.Flags;
 import seedu.todolist.exception.InvalidIdException;
 import seedu.todolist.logic.ParserUtil;
-import seedu.todolist.task.TaskList;
+import seedu.todolist.model.Config;
+import seedu.todolist.model.TaskList;
 import seedu.todolist.ui.Ui;
 
 import java.util.HashMap;
@@ -21,7 +22,7 @@ public class UnmarkTaskCommand extends Command{
     }
 
     @Override
-    public void execute(TaskList taskList, Ui ui) throws InvalidIdException {
+    public void execute(TaskList taskList, Config config, Ui ui) throws InvalidIdException {
         for (int id : idHashSet) {
             String taskString = taskList.setDone(id, false);
             ui.printUnmarkTaskMessage(taskString);
