+++ src/main/java/seedu/todolist/model/Task.java
@@ -1,4 +1,4 @@
-package seedu.todolist.task;
+package seedu.todolist.model;
 
 import seedu.todolist.constants.Formats;
 import seedu.todolist.logic.FormatterUtil;
